[c] fix compile warning (#32)
* [c] fix compile warning

* [make] add '-Wimplicit' and  '-Wconversion' flag

* [embedded_cli] avoid conversion warning by changing datatype
-> from int to size_t for counter, len, cursor
-> propagate size_t change to internal functions

---------

+++ tests/embedded_cli_test.c
@@ -148,7 +148,7 @@ static void output_putchar(void *data, char ch, bool is_last)
         // just ignore them
         if (ch >= 'A' && ch <= 'Z') {
             if (ch == 'K') // CLEAR_EOL
-                memset(&output[output_pos], 0, sizeof(output) - output_pos);
+                memset(&output[output_pos], 0, sizeof(output) - (size_t)output_pos);
             have_csi = false;
         }
     } else {
@@ -231,7 +231,7 @@ static void callback(void *data, char ch, bool is_last)
     (void)is_last;
     char *d = (char *)data;
     if (d) {
-        int len = strlen(d);
+        size_t len = strlen(d);
         if (len < MAX_OUTPUT_LEN - 1) {
             d[len] = ch;
             d[len + 1] = '\0';
